Made update_test.sh work again; made things more consistens.

+++ test/update_test.sh
@@ -24,6 +24,12 @@
 
 ########## Functions ##########################################################
 
+error() {
+  exit_status=$1; shift
+  echo $ME: $*
+  exit $exit_status
+}
+
 local_basename() {
   ##
   # Autoconf, 11.15:
@@ -59,10 +65,11 @@ ME=$(local_basename "$0")
 
 ########## Process command-line ###############################################
 
-while getopts s: opt
+while getopts B:s: opt
 do
   case $opt in
-  s) BUILD_SRC="$OPTARG" ;;
+  B) ABS_TOP_BUILDDIR="$OPTARG" ;;
+  s) TOP_SRCDIR="$OPTARG" ;;
   ?) usage ;;
   esac
 done
@@ -72,10 +79,8 @@ shift $(( OPTIND - 1 ))
 
 ########## Initialize #########################################################
 
-[ "$BUILD_SRC" ] || {
-  echo "$ME: \$BUILD_SRC not set" >&2
-  exit 2
-}
+[ "$ABS_TOP_BUILDDIR" ] || error 78 '$ABS_TOP_BUILDDIR not set'
+[ "$TOP_SRCDIR" ] || error 78 '$TOP_SRCDIR not set'
 
 [ "$TMPDIR" ] || TMPDIR=/tmp
 trap "x=$?; rm -f $TMPDIR/*_$$_* 2>/dev/null; exit $x" EXIT HUP INT TERM
@@ -90,10 +95,13 @@ DATA_DIR="$srcdir/data"
 EXPECTED_DIR="$srcdir/expected"
 ACTUAL_OUTPUT="$TMPDIR/cdecl_test_output_$$_"
 
+TEST_CONFIG="$TOP_SRCDIR/etc/config.toml"
+
 ##
-# Must put BUILD_SRC first in PATH so we get the correct version of cdecl.
+# Must put $ABS_TOP_BUILDDIR/src first in PATH so we get the correct version of
+# include-tidy.
 ##
-PATH=$BUILD_SRC:$PATH
+PATH="$ABS_TOP_BUILDDIR/src:$PATH"
 
 ##
 # Disable core dumps so we won't fill up the disk with them if a bunch of tests
